# src/multi-chat/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kuwa API",
        "description": "API definition for KuwaClient service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/create/bot": {
            "post": {
                "tags": [
                    "Bots"
                ],
                "summary": "Create a bot",
                "operationId": "f3e2fbb5bc1ec98f7cf05a51185c8f8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bot created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateBotRequest": {
                "properties": {
                    "llm_access_code": {
                        "type": "string"
                    },
                    "bot_name": {
                        "type": "string"
                    },
                    "visibility": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}